renamed screen package to window
